Migrate status plugin into a component
