+++ flexget/components/status/cli.py
@@ -9,18 +9,10 @@ from sqlalchemy import desc
 from sqlalchemy.orm.exc import NoResultFound
 
 from flexget import options
-from flexget import plugin
 from flexget.event import event
 from flexget.manager import Session
 from flexget.terminal import TerminalTable, TerminalTableError, table_parser, colorize, console
-
-try:
-    # NOTE: Importing other plugins is discouraged!
-    from flexget.plugins.operate import status as plugin_status
-except ImportError:
-    raise plugin.DependencyError(
-        issued_by=__name__, missing='status',
-    )
+from . import db
 
 
 def do_cli(manager, options):
@@ -37,12 +29,12 @@ def do_cli_task(manager, options):
     table_data = [header]
     with Session() as session:
         try:
-            task = session.query(plugin_status.StatusTask).filter(plugin_status.StatusTask.name == options.task).one()
+            task = session.query(db.StatusTask).filter(db.StatusTask.name == options.task).one()
         except NoResultFound:
             console('Task name `%s` does not exists or does not have any records' % options.task)
             return
         else:
-            query = task.executions.order_by(desc(plugin_status.TaskExecution.start))[:options.limit]
+            query = task.executions.order_by(desc(db.TaskExecution.start))[:options.limit]
             for ex in reversed(query):
                 start = ex.start.strftime('%Y-%m-%d %H:%M')
                 start = colorize('green', start) if ex.succeeded else colorize('red', start)
@@ -77,12 +69,12 @@ def do_cli_summary(manager, options):
     table_data = [header]
 
     with Session() as session:
-        for task in session.query(plugin_status.StatusTask).all():
-            ok = session.query(plugin_status.TaskExecution). \
-                filter(plugin_status.TaskExecution.task_id == task.id). \
-                filter(plugin_status.TaskExecution.succeeded == True). \
-                filter(plugin_status.TaskExecution.produced > 0). \
-                order_by(plugin_status.TaskExecution.start.desc()).first()
+        for task in session.query(db.StatusTask).all():
+            ok = session.query(db.TaskExecution). \
+                filter(db.TaskExecution.task_id == task.id). \
+                filter(db.TaskExecution.succeeded == True). \
+                filter(db.TaskExecution.produced > 0). \
+                order_by(db.TaskExecution.start.desc()).first()
 
             if ok is None:
                 duration = None
